added some admin configs sections: alerts, auth bg images, ads

+++ packages/components/src/friends/queries/set-unnote-friend.ts
@@ -1,17 +1,19 @@
 'use server';
 
 import "server-only"
-import { createClient } from '@repo/lib/utils/supabase/server.ts';
 import { getCurrentUser } from '@repo/lib/actions/get-current-user.ts';
 import { SetNote } from './set-note-friend.ts';
+import { createClient } from '@repo/lib/utils/supabase/server.ts';
 
 export async function setUnNoteFriend({
   recipient, friend_id
 }: Omit<SetNote, "note">) {
-  const supabase = createClient();
   const currentUser = await getCurrentUser();
+  if (!currentUser) return;
   
-  const { data, error, status } = await supabase
+  const api = createClient();
+  
+  const { data, error, status } = await api
   .from('friends_notes')
   .delete()
   .eq('friend_id', friend_id)
